+++ wazo_phoned/plugins/yealink/bus_consume.py
@@ -15,8 +15,17 @@ class BusEventHandler:
         bus_consumer.on_event(
             'users_services_dnd_updated', self._users_services_dnd_updated
         )
+        bus_consumer.on_event(
+            'users_forwards_busy_updated', self._users_forwards_busy_updated
+        )
 
     def _users_services_dnd_updated(self, event):
         user_uuid = event.get('user_uuid') or event.get('user_id')
         enabled = event['enabled']
         self._service.notify_dnd(user_uuid, enabled)
+
+    def _users_forwards_busy_updated(self, event):
+        user_uuid = event.get('user_uuid') or event.get('user_id')
+        enabled = event['enabled']
+        destination = event['destination']
+        self._service.notify_forward_busy(user_uuid, destination, enabled)